+++ codegen/rawscan.go
@@ -194,6 +194,8 @@ func oidKind(oid uint32) (kind, string) {
 		return kindUUIDArray, "uuid[]"
 	case 1016:
 		return kindInt8Array, "int8[]"
+	case 1007:
+		return kindInt4Array, "int4[]"
 	case 1231:
 		return kindDecimalArray, "numeric[]"
 	}
@@ -244,6 +246,8 @@ func oidGoType(oid uint32) string {
 		c.Type.Name, c.Type.Array = schema.TypeUUID, true
 	case kindInt8Array:
 		c.Type.Name, c.Type.Array = schema.TypeInt8, true
+	case kindInt4Array:
+		c.Type.Name, c.Type.Array = schema.TypeInt4, true
 	case kindDecimalArray:
 		c.Type.Name, c.Type.Array = schema.TypeNumeric, true
 	default:
@@ -348,6 +352,8 @@ func oidSchemaType(k kind) schema.Type {
 		return schema.Type{Name: schema.TypeUUID, Array: true}
 	case kindInt8Array:
 		return schema.Type{Name: schema.TypeInt8, Array: true}
+	case kindInt4Array:
+		return schema.Type{Name: schema.TypeInt4, Array: true}
 	case kindDecimalArray:
 		return schema.Type{Name: schema.TypeNumeric, Array: true}
 	}